+++ lib/src/initialization/sdk_initializer.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/widgets.dart';
-import 'package:open_cdp_flutter_sdk/src/models/config.dart';
 import 'package:open_cdp_flutter_sdk/src/implementation/sdk_implementation.dart';
 import 'package:open_cdp_flutter_sdk/src/utils/screen_tracker.dart';
 import 'package:open_cdp_flutter_sdk/src/utils/lifecycle_tracker.dart';